fix: 타입 변환

+++ src/main/java/com/example/web_chatroom/DTO/ChatCreateDTO.java
@@ -1,3 +1,3 @@
 package com.example.web_chatroom.DTO;
-public record ChatCreateDTO(String userId, String nickname, int postId, String postTitle ) {
+public record ChatCreateDTO(String userId, String nickname, Long postId, String postTitle ) {
 }
